requirements.txt
flask
azure-cognitiveservices-speech
gunicorn